Allow instruction decoding to fail per type
There are some instructions that do not permit the entire range of their
values to be decoded, such as a load/store with a register-shift offset.
As such, individual instruction types' decoding functions should be able
to return None, a feature that is now supported.

I've also re-enabled feature(bool_to_option), since it's just so
incredibly useful in concisely expressing conditional Options.
Hopefully it stabilises soon!