Made RegisterNumber a newtype
To consolidate the common tasks of extracting a register number from an
instruction and formatting it into a user-readable name, I've made
cpu::RegisterNumber a newtype.  The code outside of the tests is a bit
cleaner as a result, with most calls to bit_twiddling::select_bits()
having been factored into RegisterNumber::extract().  The slight
additional imposition of calling the r() function in the test suite is
negligible, although the refactor to use this function everywhere was a
bit tedious.

Also, I've made a new top-level module, bit_twiddling, for the standard
bit-level operations that are involved in emulating a CPU.  I figured
they would need to be extracted from cpu::instruction at some point.