Implemented decoding block data transfer instructions
Added support for decoding block data transfer instructions, which
transfer a series of registers to or from memory in a single instruction
with a number of addressing modes and register bank control options.
These were somewhat of a pain to implement, not because of the decoding
itself but because I could not find an assembler that supports the
mnemonics used by the ARMv4 documentation: presumably there is some
equivalent in more recent ISA versions, but I reverted to
hand-assembling the machine code in this case.