Changed SingleDataTransfer type to enum
The type of SingleDataTransfer instructions has been changed from a
boolean to an enum in preparation to add support for the halfword and
signed partial word transfer instructions to the same instruction type.
I am somewhat curious as to the performance impact of matching upon an
enum for each memory transfer, but I don't think that it should be any
less efficient than it would be with the boolean given sufficient
compiler optimisations.  I'll have to see if the increased code size
harms the cache performance of these very common instructions.