Implemented decoding signed and half-word transfers
Added support for decoding the "Halfword and Signed Data Transfer"
instructions as variants of the SingleDataTransfer type.  An unexpected
surprise in implementing these instructions was that they clash with a
number of the other arithmetic and data-transfer instructions; as such,
I had to re-order the decode priority table in build.rs to deviate from
the one given in the ARM documentation.  The current priorities seem to
work fine, although there may be latent errors that can only be
uncovered by starting to run some test ROMs in a bit.  Getting much
closer!