Implemented decoding Coprocessor Data Transfer
Added support for decoding the Coprocessor Data Transfer instructions,
which transfer data between coprocessor registers and memory.  In order
to clean up the code for this implementation a bit, I also added a
helper function in rgba::cpu::disassembly to format an address
represented as a (register, offset) pair, as required in several memory
transfer instruction types.