Finished ARM mode instruction decoding
With the Coprocessor Register Transfer instructions, the instruction
decoding and disassembly system is now complete; in other words, I now
have what amounts to essentially an extremely overcomplicated ARM-mode
disassembler.  Of course, I couldn't make it all the way through the
ARM documentation without spotting an error: page 4-54 contains an MRC
instruction with an opcode of 9, which is impossible given that the
opcode field for these instructions is only 3 bits wide.  I think that
the authors of the documentation likely copied the sample instructions
from the Coprocessor Data Operation instructions, which have a
4-bit-wide opcode field.